+++ R/abnormal_by_baseline.R
@@ -18,6 +18,8 @@
 #'
 #' @inheritParams argument_convention
 #' @param abnormal (`character`)\cr identifying the abnormal range level(s) in `.var`.
+#' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("abnormal_by_baseline")`
+#'   to see available statistics for this function.
 #'
 #' @note
 #' * `df` should be filtered to include only post-baseline records.