+++ R/utils_default_stats_formats_labels.R
@@ -1,13 +1,13 @@
-#' Defaults for statistical method names and their associated formats & labels
+#' Get default statistical methods and their associated formats, labels, and indent modifiers
 #'
 #' @description `r lifecycle::badge("experimental")`
 #'
 #' Utility functions to get valid statistic methods for different method groups
-#' (`.stats`) and their associated formats (`.formats`) and labels (`.labels`). This utility
-#' is used across `tern`, but some of its working principles can be seen in [analyze_vars()].
-#' See notes to understand why this is experimental.
+#' (`.stats`) and their associated formats (`.formats`), labels (`.labels`), and indent modifiers
+#' (`.indent_mods`). This utility is used across `tern`, but some of its working principles can be
+#' seen in [analyze_vars()]. See notes to understand why this is experimental.
 #'
-#' @param stats (`character`)\cr statistical methods to get defaults formats or labels for.
+#' @param stats (`character`)\cr statistical methods to get defaults for.
 #'
 #' @details
 #' Current choices for `type` are `counts` and `numeric` for [analyze_vars()] and affect `get_stats()`.
@@ -19,18 +19,18 @@
 #' @name default_stats_formats_labels
 NULL
 
-#' @describeIn default_stats_formats_labels Get defaults statistical methods for different
-#'   groups of methods.
+#' @describeIn default_stats_formats_labels Get statistics available for a given method
+#'   group (analyze function).
 #'
-#' @param method_groups (`character`)\cr indicates the group of statistical methods that
-#'   we need the defaults from. A character vector can be used to collect more than one group of statistical
-#'   methods.
-#' @param stats_in (`character`)\cr desired stats to be picked out from the selected method group.
-#' @param add_pval (`flag`)\cr should `"pval"` or `"pval_counts"` (if `method_groups` contains
+#' @param method_groups (`character`)\cr indicates the statistical method group (`tern` analyze function)
+#'   to retrieve default statistics for. A character vector can be used to specify more than one statistical
+#'   method group.
+#' @param stats_in (`character`)\cr statistics to retrieve for the selected method group.
+#' @param add_pval (`flag`)\cr should `"pval"` (or `"pval_counts"` if `method_groups` contains
 #'   `"analyze_vars_counts"`) be added to the statistical methods?
 #'
 #' @return
-#' * `get_stats()` returns a character vector with all default statistical methods.
+#' * `get_stats()` returns a `character` vector of statistical methods.
 #'
 #' @examples
 #' # analyze_vars is numeric
@@ -66,22 +66,11 @@ get_stats <- function(method_groups = "analyze_vars_numeric", stats_in = NULL, a
 
   # Loop for multiple method groups
   for (mgi in method_groups) {
-    # Main switcher
-    out_tmp <- switch(mgi,
-      "count_occurrences" = c("count", "count_fraction", "count_fraction_fixed_dp", "fraction"),
-      "summarize_num_patients" = c("unique", "nonunique", "unique_count"),
-      "analyze_vars_counts" = c("n", "count", "count_fraction", "n_blq"),
-      "analyze_vars_numeric" = c(
-        "n", "sum", "mean", "sd", "se", "mean_sd", "mean_se", "mean_ci", "mean_sei",
-        "mean_sdi", "mean_pval", "median", "mad", "median_ci", "quantiles", "iqr",
-        "range", "min", "max", "median_range", "cv", "geom_mean", "geom_mean_ci",
-        "geom_cv"
-      ),
-      "surv_time" = c("median", "median_ci", "quantiles", "range_censor", "range_event", "range"),
-      stop(
-        "The selected method group (", mgi, ") has no default statistical method."
-      )
-    )
+    out_tmp <- if (mgi %in% names(tern_default_stats)) {
+      tern_default_stats[[mgi]]
+    } else {
+      stop("The selected method group (", mgi, ") has no default statistical method.")
+    }
     out <- unique(c(out, out_tmp))
   }
 
@@ -132,14 +121,14 @@ get_stats <- function(method_groups = "analyze_vars_numeric", stats_in = NULL, a
   out
 }
 
-#' @describeIn default_stats_formats_labels Get formats from vector of statistical methods. If not
-#'   present `NULL` is returned.
+#' @describeIn default_stats_formats_labels Get formats corresponding to a list of statistics.
 #'
 #' @param formats_in (named `vector`) \cr inserted formats to replace defaults. It can be a
 #'   character vector from [formatters::list_valid_format_labels()] or a custom format function.
 #'
 #' @return
-#' * `get_formats_from_stats()` returns a named list of formats, they being a value from
+#' * `get_formats_from_stats()` returns a named vector of formats (if present in either
+#'   `tern_default_formats` or `formats_in`, otherwise `NULL`). Values can be taken from
 #'   [formatters::list_valid_format_labels()] or a custom function (e.g. [formatting_functions]).
 #'
 #' @note Formats in `tern` and `rtables` can be functions that take in the table cell value and
@@ -188,9 +177,9 @@ get_formats_from_stats <- function(stats, formats_in = NULL) {
   out
 }
 
-#' @describeIn default_stats_formats_labels Get labels from vector of statistical methods.
+#' @describeIn default_stats_formats_labels Get labels corresponding to a list of statistics.
 #'
-#' @param labels_in (named `vector`)\cr inserted labels to replace defaults.
+#' @param labels_in (named `vector` of `character`)\cr inserted labels to replace defaults.
 #' @param row_nms (`character`)\cr row names. Levels of a `factor` or `character` variable, each
 #'   of which the statistics in `.stats` will be calculated for. If this parameter is set, these
 #'   variable levels will be used as the defaults, and the names of the given custom values should
@@ -198,8 +187,8 @@ get_formats_from_stats <- function(stats, formats_in = NULL) {
 #'   variable names if rows correspond to different variables instead of levels. Defaults to `NULL`.
 #'
 #' @return
-#' * `get_labels_from_stats()` returns a named character vector of default labels (if present
-#'   otherwise `NULL`).
+#' * `get_labels_from_stats()` returns a named `character` vector of labels (if present in either
+#'   `tern_default_labels` or `labels_in`, otherwise `NULL`).
 #'
 #' @examples
 #' # Defaults labels
@@ -360,6 +349,53 @@ labels_use_control <- function(labels_default, control, labels_custom = NULL) {
 
   labels_default
 }
+
+#' @describeIn default_stats_formats_labels Named list of available statistics by method group for `tern`.
+#'
+#' @format
+#' * `tern_default_stats` is a named list of available statistics, with each element
+#'   named for their corresponding statistical method group.
+#'
+#' @export
+tern_default_stats <- list(
+  abnormal = c("fraction"),
+  abnormal_by_baseline = c("fraction"),
+  abnormal_by_marked = c("count_fraction", "count_fraction_fixed_dp"),
+  abnormal_by_worst_grade = c("count_fraction", "count_fraction_fixed_dp"),
+  abnormal_by_worst_grade_worsen = c("fraction"),
+  analyze_patients_exposure_in_cols = c("n_patients", "sum_exposure"),
+  analyze_vars_counts = c("n", "count", "count_fraction", "count_fraction_fixed_dp", "n_blq"),
+  analyze_vars_numeric = c(
+    "n", "sum", "mean", "sd", "se", "mean_sd", "mean_se", "mean_ci", "mean_sei", "mean_sdi", "mean_pval",
+    "median", "mad", "median_ci", "quantiles", "iqr", "range", "min", "max", "median_range", "cv",
+    "geom_mean", "geom_mean_ci", "geom_cv"
+  ),
+  count_cumulative = c("count_fraction", "count_fraction_fixed_dp"),
+  count_missed_doses = c("n", "count_fraction", "count_fraction_fixed_dp"),
+  count_occurrences = c("count", "count_fraction", "count_fraction_fixed_dp", "fraction"),
+  count_occurrences_by_grade = c("count_fraction", "count_fraction_fixed_dp"),
+  count_patients_with_event = c("n", "count", "count_fraction", "count_fraction_fixed_dp", "n_blq"),
+  count_patients_with_flags = c("n", "count", "count_fraction", "count_fraction_fixed_dp", "n_blq"),
+  count_values = c("n", "count", "count_fraction", "count_fraction_fixed_dp", "n_blq"),
+  coxph_pairwise = c("pvalue", "hr", "hr_ci", "n_tot", "n_tot_events"),
+  estimate_incidence_rate = c("person_years", "n_events", "rate", "rate_ci"),
+  estimate_multinomial_response = c("n_prop", "prop_ci"),
+  estimate_odds_ratio = c("or_ci", "n_tot"),
+  estimate_proportion = c("n_prop", "prop_ci"),
+  estimate_proportion_diff = c("diff", "diff_ci"),
+  summarize_ancova = c("n", "lsmean", "lsmean_diff", "lsmean_diff_ci", "pval"),
+  summarize_coxreg = c("n", "hr", "ci", "pval", "pval_inter"),
+  summarize_glm_count = c("n", "rate", "rate_ci", "rate_ratio", "rate_ratio_ci", "pval"),
+  summarize_num_patients = c("unique", "nonunique", "unique_count"),
+  summarize_patients_events_in_cols = c("unique", "all"),
+  surv_time = c("median", "median_ci", "quantiles", "range_censor", "range_event", "range"),
+  surv_timepoint = c("pt_at_risk", "event_free_rate", "rate_se", "rate_ci", "rate_diff", "rate_diff_ci", "ztest_pval"),
+  tabulate_rsp_biomarkers = c("n_tot", "n_rsp", "prop", "or", "ci", "pval"),
+  tabulate_rsp_subgroups = c("n", "n_rsp", "prop", "n_tot", "or", "ci", "pval"),
+  tabulate_survival_biomarkers = c("n_tot", "n_tot_events", "median", "hr", "ci", "pval"),
+  tabulate_survival_subgroups = c("n_tot_events", "n_events", "n_tot", "n", "median", "hr", "ci", "pval"),
+  test_proportion_diff = c("pval")
+)
 
 #' @describeIn default_stats_formats_labels Named vector of default formats for `tern`.
 #'